#39 refactor: Extract component logic into utils and services
